+++ ABLV_IngSW/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>ABLV_IngSW</name>
+	<name>ABCLV_IngSW</name>
 	<comment></comment>
 	<projects>
 	</projects>
