+++ ABLV_IngSW/src/main/java/unibg/it/ABCLV_IngSW/Algeri.java
@@ -1,4 +1,4 @@
-package unibg.it.ABLV_IngSW;
+package unibg.it.ABCLV_IngSW;
 
 public class Algeri {
 	//classe di algeri
