+++ ABLV_IngSW/src/test/java/unibg/it/ABCLV_IngSW/AppTest.java
@@ -1,4 +1,4 @@
-package unibg.it.ABLV_IngSW;
+package unibg.it.ABCLV_IngSW;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
